+++ Sources/Scaletor/Models/Accidental.swift
@@ -19,4 +19,19 @@ extension Accidental {
             self = .natural
         }
     }
+
+    internal var id: Int {
+        switch self {
+        case .doubleFlat:
+            return -2
+        case .flat:
+            return -1
+        case .natural:
+            return 0
+        case .sharp:
+            return 1
+        case .doubleSharp:
+            return 2
+        }
+    }
 }